refactor(harvest): Extract AccountField from AccountForm 🛠
