+++ packages/cozy-harvest-lib/test/components/AccountForm.spec.js
@@ -2,11 +2,7 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 
-import {
-  AccountForm,
-  AccountFields,
-  AccountField
-} from 'components/AccountForm'
+import { AccountForm, AccountFields } from 'components/AccountForm'
 
 import { KonnectorJobError } from 'helpers/konnectors'
 
@@ -272,81 +268,6 @@ describe('AccountForm', () => {
         <AccountFields manifestFields={fixtures.sanitized} t={t} />
       ).getElement()
       expect(component).toMatchSnapshot()
-    })
-  })
-
-  describe('AccountField', () => {
-    it('should render', () => {
-      const wrapper = shallow(
-        <AccountField {...fixtures.sanitized.username} name="username" t={t} />
-      )
-      const component = wrapper.dive().getElement()
-      expect(component).toMatchSnapshot()
-    })
-
-    it('render a date field', () => {
-      const wrapper = shallow(
-        <AccountField required={true} type="date" name="date" t={t} />
-      )
-      const component = wrapper.dive().getElement()
-      expect(component).toMatchSnapshot()
-    })
-
-    it('render a password field', () => {
-      const wrapper = shallow(
-        <AccountField
-          {...fixtures.sanitized.passphrase}
-          name="passphrase"
-          t={t}
-        />
-      )
-      const component = wrapper.dive().getElement()
-      expect(component).toMatchSnapshot()
-    })
-
-    it('render a dropdown field', () => {
-      const options = [
-        { label: 'Option 1', value: 'option1' },
-        { label: 'Option 2', value: 'option2' }
-      ]
-
-      const wrapper = shallow(
-        <AccountField
-          name="multiple"
-          options={options}
-          required={true}
-          t={t}
-          type="dropdown"
-        />
-      )
-      const component = wrapper.dive().getElement()
-      expect(component).toMatchSnapshot()
-    })
-
-    it('uses predefined label', () => {
-      const wrapper = shallow(
-        <AccountField
-          label="login"
-          name="username"
-          required={true}
-          type="text"
-          t={t}
-        />
-      )
-      expect(wrapper.props().label).toBe('fields.login.label')
-    })
-
-    it('ignores invalid predefined label', () => {
-      const wrapper = shallow(
-        <AccountField
-          label="foo"
-          name="username"
-          required={true}
-          type="text"
-          t={t}
-        />
-      )
-      expect(wrapper.props().label).toBe('fields.username.label')
     })
   })
 })
